% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorBestVariants.r
\name{.get_omim_dt}
\alias{.get_omim_dt}
\title{Load and cache the OMIM phenotype table}
\usage{
.get_omim_dt(package = "ClinPrior")
}
\description{
Load and cache the OMIM phenotype table
}
\keyword{internal}
