% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxentScanClinPriorVcfR.r
\name{MaxentScanClinPriorVcfR}
\alias{MaxentScanClinPriorVcfR}
\title{Title}
\usage{
MaxentScanClinPriorVcfR(
  matrixVariants,
  assembly = "assembly37",
  Threshold = 0.2
)
}
\arguments{
\item{matrixVariants}{}
}
\value{

}
\description{
Title
}
