% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorBestVariantsVcfR.r
\name{priorBestVariantVcfR}
\alias{priorBestVariantVcfR}
\title{Title}
\usage{
priorBestVariantVcfR(
  vcfRObject = variants,
  sampleVCF = sampleName,
  filter = "",
  isCodingVar = TRUE,
  frequenceAR = 0.01,
  frequenceAD = 0,
  qualityThreshold = 0,
  GlobalPhenotypicScore = PriorPhenoScore,
  assembly = "assembly37",
  processors = 4
)
}
\arguments{
\item{isCodingVar}{}

\item{processors}{}
}
\value{



}
\description{
Title
}
\examples{
Title

}
