% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinScore.r
\name{proteinScore}
\alias{proteinScore}
\title{Phenotypic score for each gene based on the description of the patient in HPO terms}
\usage{
proteinScore(HPOpatient)
}
\arguments{
\item{HPOpatient}{vector with the patient HPO terms}
}
\value{
vector with the gene scores
}
\description{
Phenotypic score for each gene based on the description of the patient in HPO terms
}
\examples{
HPOpatient = c("HP:0004481","HP:0002376","HP:0001257","HP:0001250","HP:0000238","HP:0002922","HP:0000365")
Y<-proteinScore(HPOpatient)
}
